@@ -1,6 +1,6 @@
 {
   "name": "@weni/unnnic-system",
-  "version": "3.1.1",
+  "version": "3.4.0",
   "type": "commonjs",
   "files": [
     "dist",
@@ -30,14 +30,15 @@
     "build": "vite build",
     "build:types": "vue-tsc && vite build",
     "preview": "vite preview",
-    "lint": "eslint --ext .js,.vue,.ts --ignore-path .gitignore --fix src",
+    "lint": "eslint src/**/*.{js,vue,ts} --fix",
     "format": "prettier --write src/",
     "type-check": "vue-tsc --noEmit",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "test": "vitest",
     "test:ui": "vitest --ui",
-    "coverage": "vitest --coverage"
+    "coverage": "vitest --coverage",
+    "style-dictionary:generate": "style-dictionary build --config styleDictionary.config.js"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
@@ -50,14 +51,16 @@
     "vue": "^3.4.8"
   },
   "dependencies": {
-    "@emoji-mart/data": "^1.1.2",
     "@vueuse/components": "^10.4.1",
-    "emoji-mart": "^5.5.2",
+    "emoji-mart-vue-fast": "^15.0.5",
     "lodash": "^4.17.21",
+    "lucide-vue-next": "^0.546.0",
     "mime": "^4.0.1",
     "mitt": "^3.0.1",
     "moment": "^2.30.1",
+    "reka-ui": "^2.6.0",
     "remark-gfm": "^4.0.0",
+    "tw-animate-css": "^1.4.0",
     "vue": "^3.4.8",
     "vue-i18n": "9",
     "vue-the-mask": "^0.11.1"
@@ -80,18 +83,28 @@
     "@vitest/coverage-istanbul": "^1.6.0",
     "@vitest/coverage-v8": "^1.6.0",
     "@vitest/ui": "^1.6.0",
-    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "@vue/eslint-config-typescript": "^14.6.0",
     "@vue/test-utils": "^2.4.6",
-    "eslint": "^8.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
+    "@weni/eslint-config": "^2.0.0",
+    "ajv": "^8.17.1",
+    "autoprefixer": "^10.4.21",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "eslint": "^9.37.0",
     "eslint-plugin-storybook": "^0.8.0",
-    "eslint-plugin-vue": "^9.25.0",
+    "eslint-plugin-vue": "10.4.0",
     "jsdom": "^24.1.1",
+    "postcss": "^8.5.6",
+    "prettier": "3.6.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "^1.63.0",
     "storybook": "^8.0.10",
+    "style-dictionary": "^5.0.1",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss": "^3.4.18",
+    "tailwindcss-animate": "^1.0.7",
     "typescript": "^5.9.2",
     "vite": "4.3.5",
     "vite-plugin-dts": "^4.5.4",
