@@ -13,8 +13,8 @@
     data-testid="material-icon"
     translate="no"
     @click="onClick"
-    @mousedown="$emit('mousedown')"
-    @mouseup="$emit('mouseup')"
+    @mousedown="(event) => $emit('mousedown', event)"
+    @mouseup="(event) => $emit('mouseup', event)"
   >
     {{ materialSymbolsName }}
   </span>
@@ -31,87 +31,57 @@
     ]"
     data-testid="old-map-icons"
     @click="onClick"
-    @mousedown="$emit('mousedown')"
-    @mouseup="$emit('mouseup')"
+    @mousedown="(event) => $emit('mousedown', event)"
+    @mouseup="(event) => $emit('mouseup', event)"
   />
 </template>
 
-<script>
+<script setup lang="ts">
+import { computed } from 'vue';
 import icons from '../utils/icons';
 import OldIconsMap from './Icon/OldIconsMap.json';
-/* eslint-disable vue/multi-word-component-names */
-export default {
-  name: 'Icon',
-  props: {
-    filled: {
-      type: Boolean,
-    },
-    next: {
-      type: Boolean,
-    },
-    icon: {
-      type: String,
-      default: null,
-    },
-    clickable: {
-      type: Boolean,
-      default: false,
-    },
-    size: {
-      type: String,
-      default: 'md',
-      validator(value) {
-        return (
-          [
-            'nano',
-            'xs',
-            'sm',
-            'ant',
-            'md',
-            'lg',
-            'xl',
-            'avatar-lg',
-            'avatar-md',
-            'avatar-sm',
-            'avatar-xs',
-            'avatar-nano',
-          ].indexOf(value) !== -1
-        );
-      },
-    },
-    lineHeight: {
-      type: String,
-      default: null,
-      validator(value) {
-        return !value || ['sm', 'md', 'lg'].indexOf(value) !== -1;
-      },
-    },
-    scheme: {
-      type: String,
-      default: 'neutral-darkest',
-    },
-  },
-  emits: ['click', 'mousedown', 'mouseup'],
-  computed: {
-    svg() {
-      return icons[this.icon];
-    },
+import type { IconProps, IconSize, LineHeight } from './Icon/types';
+import type { SchemeColor } from '@/types/scheme-colors';
 
-    materialSymbolsName() {
-      if (Object.keys(icons).includes(this.icon) && !this.next) {
-        return null;
-      }
+defineOptions({
+  name: 'UnnnicIcon',
+});
 
-      return OldIconsMap[this.icon] || this.icon;
-    },
-  },
-  mounted() {},
-  methods: {
-    onClick($event) {
-      if (!this.clickable) return;
-      this.$emit('click', $event);
-    },
-  },
+export type { IconProps, IconSize, LineHeight, SchemeColor };
+
+const props = withDefaults(defineProps<IconProps>(), {
+  filled: false,
+  next: false,
+  icon: null,
+  clickable: false,
+  size: 'md',
+  lineHeight: null,
+  scheme: 'neutral-darkest',
+});
+
+const emit = defineEmits<{
+  click: [event: Event];
+  mousedown: [event: Event];
+  mouseup: [event: Event];
+}>();
+
+const svg = computed(() => {
+  return icons[props.icon as string];
+});
+
+const materialSymbolsName = computed(() => {
+  if (Object.keys(icons).includes(props.icon as string) && !props.next) {
+    return null;
+  }
+
+  return (
+    (OldIconsMap as Record<string, string>)[props.icon as string] || props.icon
+  );
+});
+
+const onClick = (event: Event) => {
+  if (!props.clickable) return;
+  emit('click', event);
 };
 </script>
 
@@ -127,95 +97,6 @@ export default {
 
 <style lang="scss">
 @use '@/assets/scss/unnnic' as *;
-
-$scheme-colors:
-  'background-solo' $unnnic-color-background-solo,
-  'background-sky' $unnnic-color-background-sky,
-  'background-grass' $unnnic-color-background-grass,
-  'background-carpet' $unnnic-color-background-carpet,
-  'background-snow' $unnnic-color-background-snow,
-  'background-lightest' $unnnic-color-background-lightest,
-  'background-white' $unnnic-color-background-white,
-  'neutral-black' $unnnic-color-neutral-black,
-  'neutral-darkest' $unnnic-color-neutral-darkest,
-  'neutral-dark' $unnnic-color-neutral-dark,
-  'neutral-cloudy' $unnnic-color-neutral-cloudy,
-  'neutral-cleanest' $unnnic-color-neutral-cleanest,
-  'neutral-clean' $unnnic-color-neutral-clean,
-  'neutral-soft' $unnnic-color-neutral-soft,
-  'neutral-lightest' $unnnic-color-neutral-lightest,
-  'neutral-light' $unnnic-color-neutral-light,
-  'neutral-white' $unnnic-color-neutral-white,
-  'neutral-snow' $unnnic-color-neutral-snow,
-  'feedback-red' $unnnic-color-feedback-red,
-  'feedback-green' $unnnic-color-feedback-green,
-  'feedback-yellow' $unnnic-color-feedback-yellow,
-  'feedback-blue' $unnnic-color-feedback-blue,
-  'feedback-grey' $unnnic-color-feedback-grey,
-  'aux-blue' $unnnic-color-aux-blue,
-  'aux-purple' $unnnic-color-aux-purple,
-  'aux-orange' $unnnic-color-aux-orange,
-  'aux-lemon' $unnnic-color-aux-lemon,
-  'aux-pink' $unnnic-color-aux-pink,
-  'aux-baby-blue' $unnnic-color-aux-baby-blue,
-  'aux-baby-yellow' $unnnic-color-aux-baby-yellow,
-  'aux-baby-red' $unnnic-color-aux-baby-red,
-  'aux-baby-green' $unnnic-color-aux-baby-green,
-  'aux-baby-pink' $unnnic-color-aux-baby-pink,
-  'aux-green-100' $unnnic-color-aux-green-100,
-  'aux-green-300' $unnnic-color-aux-green-300,
-  'aux-green-500' $unnnic-color-aux-green-500,
-  'aux-green-700' $unnnic-color-aux-green-700,
-  'aux-green-900' $unnnic-color-aux-green-900,
-  'aux-blue-100' $unnnic-color-aux-blue-100,
-  'aux-blue-300' $unnnic-color-aux-blue-300,
-  'aux-blue-500' $unnnic-color-aux-blue-500,
-  'aux-blue-700' $unnnic-color-aux-blue-700,
-  'aux-blue-900' $unnnic-color-aux-blue-900,
-  'aux-purple-100' $unnnic-color-aux-purple-100,
-  'aux-purple-300' $unnnic-color-aux-purple-300,
-  'aux-purple-500' $unnnic-color-aux-purple-500,
-  'aux-purple-700' $unnnic-color-aux-purple-700,
-  'aux-purple-900' $unnnic-color-aux-purple-900,
-  'aux-red-100' $unnnic-color-aux-red-100,
-  'aux-red-300' $unnnic-color-aux-red-300,
-  'aux-red-500' $unnnic-color-aux-red-500,
-  'aux-red-700' $unnnic-color-aux-red-700,
-  'aux-red-900' $unnnic-color-aux-red-900,
-  'aux-orange-100' $unnnic-color-aux-orange-100,
-  'aux-orange-300' $unnnic-color-aux-orange-300,
-  'aux-orange-500' $unnnic-color-aux-orange-500,
-  'aux-orange-700' $unnnic-color-aux-orange-700,
-  'aux-orange-900' $unnnic-color-aux-orange-900,
-  'aux-yellow-100' $unnnic-color-aux-yellow-100,
-  'aux-yellow-300' $unnnic-color-aux-yellow-300,
-  'aux-yellow-500' $unnnic-color-aux-yellow-500,
-  'aux-yellow-700' $unnnic-color-aux-yellow-700,
-  'aux-yellow-900' $unnnic-color-aux-yellow-900,
-  'floweditor-blue' $unnnic-color-floweditor-blue,
-  'floweditor-purple' $unnnic-color-floweditor-purple,
-  'floweditor-wine' $unnnic-color-floweditor-wine,
-  'floweditor-orange' $unnnic-color-floweditor-orange,
-  'floweditor-pink' $unnnic-color-floweditor-pink,
-  'floweditor-turquoise' $unnnic-color-floweditor-turquoise,
-  'floweditor-green' $unnnic-color-floweditor-green,
-  'weni-50' $unnnic-color-weni-50,
-  'weni-100' $unnnic-color-weni-100,
-  'weni-200' $unnnic-color-weni-200,
-  'weni-300' $unnnic-color-weni-300,
-  'weni-400' $unnnic-color-weni-400,
-  'weni-500' $unnnic-color-weni-500,
-  'weni-600' $unnnic-color-weni-600,
-  'weni-700' $unnnic-color-weni-700,
-  'weni-800' $unnnic-color-weni-800,
-  'weni-900' $unnnic-color-weni-900,
-  'weni-950' $unnnic-color-weni-950,
-  'brand-weni' $unnnic-color-brand-weni,
-  'brand-weni-dark' $unnnic-color-brand-weni-dark,
-  'brand-weni-soft' $unnnic-color-brand-weni-soft,
-  'brand-sec-dark' $unnnic-color-brand-sec-dark,
-  'brand-sec-soft' $unnnic-color-brand-sec-soft,
-  'brand-sec' $unnnic-color-brand-sec;
 
 $icon-sizes:
   'xl' $unnnic-icon-size-xl,
@@ -258,6 +139,7 @@ $icon-sizes:
   white-space: nowrap;
   word-wrap: normal;
   direction: ltr;
+  font-feature-settings: 'liga';
   -webkit-font-feature-settings: 'liga';
   -webkit-font-smoothing: antialiased;
 
@@ -265,7 +147,7 @@ $icon-sizes:
     font-family: 'Material Symbols Rounded Filled';
   }
 
-  @each $name, $color in $scheme-colors {
+  @each $name, $color in $unnnic-scheme-colors {
     &.unnnic-icon-scheme--#{$name} {
       color: $color;
     }
