@@ -1,63 +1,125 @@
 <template>
   <div
-    ref="pickerContainer"
+    ref="emojiPickerRef"
     :class="['emoji-picker', `emoji-picker--${position}`]"
     @click.stop
     @keypress.enter.stop
-  />
+  >
+    <Picker
+      :data="emojiIndex"
+      set="apple"
+      theme="light"
+      emoji="point_up"
+      :title="title"
+      :showPreview="true"
+      showSkinTones
+      :search="true"
+      navPosition="bottom"
+      noResultsEmoji="cry"
+      :maxFrequentRows="3"
+      :i18n="translations"
+      :color="accentColor"
+      @select="onEmojiSelect"
+    />
+  </div>
 </template>
 
-<script>
-import i18n from '@emoji-mart/data/i18n/pt.json';
-import data from '@emoji-mart/data/sets/14/apple.json';
-import { Picker } from 'emoji-mart/';
+<script setup lang="ts">
+import { computed, ref, onMounted, onUnmounted } from 'vue';
+import { get } from 'lodash';
+import i18n from '../../utils/plugins/i18n';
+import { Picker, EmojiIndex } from 'emoji-mart-vue-fast/src';
+import data from 'emoji-mart-vue-fast/data/all.json';
+import 'emoji-mart-vue-fast/css/emoji-mart.css';
+import UnnnicI18n from '../../mixins/i18n';
 
-export default {
-  name: 'UnnnicEmojiPicker',
-  props: {
-    returnName: {
-      type: Boolean,
-      default: false,
-    },
-    position: {
-      type: String,
-      default: 'top',
-      validator: (position) => ['top', 'bottom'].includes(position),
-    },
+defineOptions({
+  mixins: [UnnnicI18n],
+});
+
+interface Emoji {
+  id: string;
+  native: string;
+}
+
+export interface Props {
+  returnName?: boolean;
+  position?: 'top' | 'bottom';
+  locale?: string;
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  returnName: false,
+  position: 'top',
+  locale: 'pt-br',
+});
+
+const emit = defineEmits<{
+  close: [];
+  emojiSelected: [emoji: string];
+}>();
+
+const emojiPickerRef = ref<HTMLElement>();
+const emojiIndex = computed(() => new EmojiIndex(data));
+
+const accentColor = '#00A49F'; // $unnnic-color-weni-600
+
+const currentLocale = computed(() => props.locale || 'pt-br');
+
+const translation = (key: string) => {
+  const messages: Record<string, unknown> = i18n.global.messages as Record<
+    string,
+    unknown
+  >;
+  const loc = currentLocale.value;
+  const localeMsgs =
+    messages?.[loc] ||
+    messages?.[loc?.toLowerCase?.()] ||
+    messages?.[loc?.toUpperCase?.()];
+  const enMsgs = messages?.['en'];
+  return get(localeMsgs, key) || get(enMsgs, key) || key;
+};
+
+const title = computed(() => translation('emoji_picker.title'));
+
+const translations = computed(() => ({
+  search: translation('emoji_picker.search'),
+  notfound: translation('emoji_picker.notfound'),
+  categories: {
+    search: translation('emoji_picker.categories.search'),
+    recent: translation('emoji_picker.categories.recent'),
+    smileys: translation('emoji_picker.categories.smileys'),
+    people: translation('emoji_picker.categories.people'),
+    nature: translation('emoji_picker.categories.nature'),
+    foods: translation('emoji_picker.categories.foods'),
+    activity: translation('emoji_picker.categories.activity'),
+    places: translation('emoji_picker.categories.places'),
+    objects: translation('emoji_picker.categories.objects'),
+    symbols: translation('emoji_picker.categories.symbols'),
+    flags: translation('emoji_picker.categories.flags'),
+    custom: translation('emoji_picker.categories.custom'),
   },
-  emits: ['close', 'emojiSelected'],
-  computed: {
-    emojiPickerPreferences() {
-      return {
-        data,
-        set: 'apple',
-        theme: 'light',
-        previewPosition: 'none',
-        searchPosition: 'none',
-        navPosition: 'bottom',
-        noResultsEmoji: 'cry',
-        maxFrequentRows: 3,
-      };
-    },
-  },
-  mounted() {
-    this.initPicker();
-  },
-  methods: {
-    initPicker() {
-      // eslint-disable-next-line no-new
-      new Picker({
-        i18n,
-        parent: this.$refs.pickerContainer,
-        onEmojiSelect: this.onEmojiSelect,
-        onClickOutside: () => this.$emit('close'),
-        ...this.emojiPickerPreferences,
-      });
-    },
-    onEmojiSelect(emoji) {
-      this.$emit('emojiSelected', this.returnName ? emoji.id : emoji.native);
-    },
-  },
+}));
+
+const handleClickOutside = (event: Event) => {
+  if (
+    emojiPickerRef.value &&
+    !emojiPickerRef.value.contains(event.target as Node)
+  ) {
+    emit('close');
+  }
+};
+
+onMounted(() => {
+  document.addEventListener('click', handleClickOutside);
+});
+
+onUnmounted(() => {
+  document.removeEventListener('click', handleClickOutside);
+});
+
+const onEmojiSelect = (emoji: Emoji) => {
+  emit('emojiSelected', props.returnName ? emoji.id : emoji.native);
 };
 </script>
 
@@ -78,16 +140,45 @@ export default {
     animation-name: slideInDown;
   }
 
-  :deep(em-emoji-picker) {
-    // Most variables don't work here
-    --border-radius: 16px;
-    --font-family: Lato, sans-serif; // $unnnic-font-family
-    --rgb-accent: 0, 164, 159; // $unnnic-color-weni-600
-    --rgb-color: 59, 65, 77; // $unnnic-color-neutral-darkest
-    --color-border: rgb(244, 246, 248);
-    --shadow: 0px 4px 8px 0px rgba(0, 0, 0, 0.1);
-
+  :deep(.emoji-mart) {
+    border-radius: 16px;
+    font-family: Lato, sans-serif; // $unnnic-font-family
+    border: 1px solid rgb(244, 246, 248);
+    box-shadow: 0px 4px 8px 0px rgba(0, 0, 0, 0.1);
     cursor: default;
+    color: #3b4151; // $unnnic-color-neutral-darkest
+  }
+
+  :deep(.emoji-mart-emoji) {
+    cursor: pointer;
+  }
+
+  :deep(.emoji-mart-category .emoji-mart-emoji span) {
+    cursor: pointer;
+  }
+
+  :deep(.emoji-mart-anchor) {
+    cursor: pointer;
+    color: #858585;
+  }
+
+  :deep(.emoji-mart-anchor:hover),
+  :deep(.emoji-mart-anchor-selected) {
+    color: #00a49f; // $unnnic-color-weni-600
+  }
+
+  :deep(.emoji-mart-anchor-bar) {
+    background-color: #00a49f; // $unnnic-color-weni-600
+  }
+
+  :deep(.emoji-type-image.emoji-set-apple) {
+    background-image: url('./apple-64.png') !important;
+  }
+
+  :deep(.emoji-type-image.emoji-set-facebook),
+  :deep(.emoji-type-image.emoji-set-google),
+  :deep(.emoji-type-image.emoji-set-twitter) {
+    background-image: none !important;
   }
 }
 
