@@ -214,7 +214,12 @@ export default {
     },
   },
 
-  emits: ['update:searchValue', 'onChange', 'update:modelValue'],
+  emits: [
+    'update:searchValue',
+    'onChange',
+    'update:modelValue',
+    'onActiveChange',
+  ],
 
   data() {
     return {
@@ -303,6 +308,8 @@ export default {
     active(newValue) {
       this.$refs['dropdown-skeleton'].calculatePosition();
 
+      this.$emit('onActiveChange', newValue);
+
       this.$nextTick(() => {
         if (newValue && !this.multiple) {
           const activeOptionIndex = this.getOptionIndex('active');
@@ -381,11 +388,18 @@ export default {
           return;
         }
 
-        if(this.multiple && option.disableRemove && this.optionIsSelected(option)) {
+        if (
+          this.multiple &&
+          option.disableRemove &&
+          this.optionIsSelected(option)
+        ) {
           return;
         }
 
-        if (this.multipleLimit && this.modelValue.length >= this.multipleLimit) {
+        if (
+          this.multipleLimit &&
+          this.modelValue.length >= this.multipleLimit
+        ) {
           return;
         }
 
@@ -559,7 +573,6 @@ export default {
         const focusedOptionIndex = this.getOptionIndex('focused');
         let newIndex;
 
-        // eslint-disable-next-line default-case
         switch (key) {
           case 'Escape':
             this.active = false;
