@@ -1,96 +1,97 @@
-import type { Component } from "vue";
-import { unnnicFontSize as fontSize } from "./config";
-import formElement from "./FormElement/FormElement.vue";
-import input from "./Input/Input.vue";
-import inputNext from "./InputNext/InputNext.vue";
-import inputDatePicker from "./InputDatePicker/InputDatePicker.vue";
-import button from "./Button/Button.vue";
-import buttonIcon from "./Button/ButtonIcon.vue";
-import sidebar from "./Sidebar/index.vue";
-import sidebarItem from "./Sidebar/SidebarItem.vue";
-import table from "./Table/Table.vue";
-import tableRow from "./Table/TableRow.vue";
-import dropdown from "./Dropdown/Dropdown.vue";
-import dropdownItem from "./Dropdown/DropdownItem.vue";
-import avatarIcon from "./AvatarIcon/AvatarIcon.vue";
-import icon from "./Icon.vue";
-import iconLoading from "./IconLoading/IconLoading.vue";
-import toolTip from "./ToolTip/ToolTip.vue";
-import card from "./Card/Card.vue";
-import cardSimple from "./Card/SimpleCard.vue";
-import cardCompany from "./Card/CardCompany.vue";
-import cardData from "./Card/CardData.vue";
-import cardImage from "./CardImage/CardImage.vue";
+import type { Component } from 'vue';
+import { unnnicFontSize as fontSize } from './config';
+import formElement from './FormElement/FormElement.vue';
+import input from './Input/Input.vue';
+import inputNext from './InputNext/InputNext.vue';
+import inputDatePicker from './InputDatePicker/InputDatePicker.vue';
+import button from './Button/Button.vue';
+import buttonIcon from './Button/ButtonIcon.vue';
+import sidebar from './Sidebar/index.vue';
+import sidebarItem from './Sidebar/SidebarItem.vue';
+import table from './Table/Table.vue';
+import tableRow from './Table/TableRow.vue';
+import dropdown from './Dropdown/Dropdown.vue';
+import dropdownItem from './Dropdown/DropdownItem.vue';
+import avatarIcon from './AvatarIcon/AvatarIcon.vue';
+import icon from './Icon.vue';
+import iconLoading from './IconLoading/IconLoading.vue';
+import toolTip from './ToolTip/ToolTip.vue';
+import card from './Card/Card.vue';
+import cardSimple from './Card/SimpleCard.vue';
+import cardCompany from './Card/CardCompany.vue';
+import cardData from './Card/CardData.vue';
+import cardImage from './CardImage/CardImage.vue';
 // import cardFlow from './CardFlow/CardFlow.vue';
-import cardProject from "./CardProject/CardProject.vue";
-import cardInformation from "./CardInformation/CardInformation.vue";
-import checkbox from "./Checkbox/Checkbox.vue";
-import collapse from "./Collapse/Collapse.vue";
-import radio from "./Radio/Radio.vue";
-import languageSelect from "./Dropdown/LanguageSelect.vue";
-import modal from "./Modal/Modal.vue";
-import modalUpload from "./ModalUpload/ModalUpload.vue";
-import { callAlert, callModal } from "../utils/call";
-import selectSmart from "./SelectSmart/SelectSmart.vue";
+import cardProject from './CardProject/CardProject.vue';
+import cardInformation from './CardInformation/CardInformation.vue';
+import checkbox from './Checkbox/Checkbox.vue';
+import collapse from './Collapse/Collapse.vue';
+import radio from './Radio/Radio.vue';
+import languageSelect from './Dropdown/LanguageSelect.vue';
+import modal from './Modal/Modal.vue';
+import modalUpload from './ModalUpload/ModalUpload.vue';
+import { callAlert, callModal } from '../utils/call';
+import selectSmart from './SelectSmart/SelectSmart.vue';
 // import select from './Select/Select.vue';
-import selectItem from "./Select/SelectItem.vue";
+import selectItem from './Select/SelectItem.vue';
 // import selectListItem from './SelectListItem/SelectListItem.vue';
-import multiSelect from "./MultiSelect/MultiSelect.vue";
-import alert from "./Alert/Alert.vue";
+import multiSelect from './MultiSelect/MultiSelect.vue';
+import alert from './Alert/Alert.vue';
 // import autocomplete from './Input/Autocomplete.vue';
 // import autocompleteSelect from './AutocompleteSelect/AutocompleteSelect.vue';
-import tag from "./Tag/Tag.vue";
-import accordion from "./Accordion/Accordion.vue";
-import indicator from "./Indicator/Indicator.vue";
-import skeletonLoading from "./SkeletonLoading/SkeletonLoading.vue"; // TODO: Unbreak this component
-import carousel from "./Carousel/Carousel.vue";
-import label from "./Label/Label.vue";
-import tab from "./Tab/Tab.vue";
-import tabsExpanded from "./TabsExpanded/TabsExpanded.vue";
-import banner from "./Banner/Banner.vue";
-import comment from "./Comment/Comment.vue";
-import datePicker from "./DatePicker/DatePicker.vue";
-import breadcrumb from "./Breadcrumb/Breadcrumb.vue";
-import Switch from "./Switch/Switch.vue";
-import Slider from "./Slider/Slider.vue";
-import DataArea from "./DataArea/DataArea.vue";
-import Pagination from "./Pagination/Pagination.vue";
-import DropArea from "./DropArea/DropArea.vue";
-import UploadArea from "./UploadArea/UploadArea.vue";
-import ImportCard from "./ImportCard/ImportCard.vue";
-import DateFilter from "./DateFilter/DateFilter.vue";
-import ChatText from "./ChatText/ChatText.vue";
-import TextArea from "./TextArea/TextArea.vue";
-import CardNumber from "./CardNumber/CardNumber.vue";
-import chartRainbow from "./ChartRainbow/ChartRainbow.vue";
-import chartBar from "./ChartBar/ChartBar.vue";
-import chartLine from "./ChartLine/ChartLine.vue";
-import moodRating from "./MoodRating/MoodRating.vue";
-import starRating from "./StarRating/StarRating.vue";
-import audioRecorder from "./AudioRecorder/AudioRecorder.vue";
-import circleProgressBar from "./CircleProgressBar/CircleProgressBar.vue";
-import progressBar from "./ProgressBar/ProgressBar.vue";
-import ChatsContact from "./ChatsContact/ChatsContact.vue";
-import ChatsDashboardTagLive from "./ChatsDashboardTagLive/ChatsDashboardTagLive.vue";
-import ChatsHeader from "./ChatsHeader/ChatsHeader.vue";
-import ChatsMessage from "./ChatsMessage/ChatsMessage.vue";
-import ReplyMessage from "./ChatsMessage/ReplyMessage.vue";
-import ChatsNavbar from "./ChatsNavbar/ChatsNavbar.vue";
-import ChatsUserAvatar from "./ChatsUserAvatar/ChatsUserAvatar.vue";
-import ChartMultiLine from "./ChartMultiLine/ChartMultiLine.vue";
-import EmojiPicker from "./EmojiPicker/EmojiPicker.vue";
-import ChartFunnel from "./ChartFunnel/ChartFunnel.vue";
-import Disclaimer from "./Disclaimer/Disclaimer.vue";
-import Drawer from "./Drawer/Drawer.vue";
-import TableNext from "./TableNext/TableNext.vue";
-import ModalNext from "./ModalNext/ModalNext.vue";
-import ModalDialog from "./ModalDialog/ModalDialog.vue";
-import Tour from "./Tour/Tour.vue";
-import Navigator from "./Navigator/index.vue";
-import SelectTime from "./SelectTime/index.vue";
-import TemplatePreview from "./TemplatePreview/TemplatePreview.vue";
-import TemplatePreviewModal from "./TemplatePreview/TemplatePreviewModal.vue";
-import DataTable from "./DataTable/index.vue";
+import tag from './Tag/Tag.vue';
+import accordion from './Accordion/Accordion.vue';
+import indicator from './Indicator/Indicator.vue';
+import skeletonLoading from './SkeletonLoading/SkeletonLoading.vue'; // TODO: Unbreak this component
+import carousel from './Carousel/Carousel.vue';
+import label from './Label/Label.vue';
+import tab from './Tab/Tab.vue';
+import tabsExpanded from './TabsExpanded/TabsExpanded.vue';
+import banner from './Banner/Banner.vue';
+import comment from './Comment/Comment.vue';
+import datePicker from './DatePicker/DatePicker.vue';
+import breadcrumb from './Breadcrumb/Breadcrumb.vue';
+import Switch from './Switch/Switch.vue';
+import Slider from './Slider/Slider.vue';
+import DataArea from './DataArea/DataArea.vue';
+import Pagination from './Pagination/Pagination.vue';
+import DropArea from './DropArea/DropArea.vue';
+import UploadArea from './UploadArea/UploadArea.vue';
+import ImportCard from './ImportCard/ImportCard.vue';
+import DateFilter from './DateFilter/DateFilter.vue';
+import ChatText from './ChatText/ChatText.vue';
+import TextArea from './TextArea/TextArea.vue';
+import CardNumber from './CardNumber/CardNumber.vue';
+import chartRainbow from './ChartRainbow/ChartRainbow.vue';
+import chartBar from './ChartBar/ChartBar.vue';
+import chartLine from './ChartLine/ChartLine.vue';
+import moodRating from './MoodRating/MoodRating.vue';
+import starRating from './StarRating/StarRating.vue';
+import audioRecorder from './AudioRecorder/AudioRecorder.vue';
+import circleProgressBar from './CircleProgressBar/CircleProgressBar.vue';
+import progressBar from './ProgressBar/ProgressBar.vue';
+import ChatsContact from './ChatsContact/ChatsContact.vue';
+import ChatsDashboardTagLive from './ChatsDashboardTagLive/ChatsDashboardTagLive.vue';
+import ChatsHeader from './ChatsHeader/ChatsHeader.vue';
+import ChatsMessage from './ChatsMessage/ChatsMessage.vue';
+import ReplyMessage from './ChatsMessage/ReplyMessage.vue';
+import ChatsNavbar from './ChatsNavbar/ChatsNavbar.vue';
+import ChatsUserAvatar from './ChatsUserAvatar/ChatsUserAvatar.vue';
+import ChartMultiLine from './ChartMultiLine/ChartMultiLine.vue';
+import EmojiPicker from './EmojiPicker/EmojiPicker.vue';
+import ChartFunnel from './ChartFunnel/ChartFunnel.vue';
+import Disclaimer from './Disclaimer/Disclaimer.vue';
+import Drawer from './Drawer/Drawer.vue';
+import TableNext from './TableNext/TableNext.vue';
+import ModalNext from './ModalNext/ModalNext.vue';
+import ModalDialog from './ModalDialog/ModalDialog.vue';
+import Tour from './Tour/Tour.vue';
+import Navigator from './Navigator/index.vue';
+import SelectTime from './SelectTime/index.vue';
+import DataTable from './DataTable/index.vue';
+import Chip from './Chip/Chip.vue';
+import TemplatePreview from './TemplatePreview/TemplatePreview.vue';
+import TemplatePreviewModal from './TemplatePreview/TemplatePreviewModal.vue';
 
 type VueComponent = Component;
 
@@ -192,6 +193,7 @@ export const components: ComponentsMap = {
   unnnicTemplatePreview: TemplatePreview,
   unnnicTemplatePreviewModal: TemplatePreviewModal,
   unnnicDataTable: DataTable,
+  unnnicChip: Chip,
 };
 
 export const unnnicFontSize = fontSize;
@@ -286,5 +288,102 @@ export const unnnicTour = Tour;
 export const unnnicNavigator = Navigator;
 export const unnnicDataTable = DataTable as VueComponent;
 export const unnnicSelectTime = SelectTime as VueComponent;
+export const unnnicChip = Chip;
 export const unnnicTemplatePreview = TemplatePreview as VueComponent;
 export const unnnicTemplatePreviewModal = TemplatePreviewModal as VueComponent;
+
+export const UnnnicFontSize = fontSize;
+export const UnnnicFormElement = formElement;
+export const UnnnicInput = input;
+export const UnnnicInputNext = inputNext;
+export const UnnnicInputDatePicker = inputDatePicker;
+export const UnnnicButton = button;
+export const UnnnicButtonIcon = buttonIcon;
+export const UnnnicSidebar = sidebar;
+export const UnnnicSidebarItem = sidebarItem;
+export const UnnnicTable = table;
+export const UnnnicTableRow = tableRow;
+export const UnnnicDropdown = dropdown as VueComponent;
+export const UnnnicDropdownItem = dropdownItem;
+export const UnnnicAvatarIcon = avatarIcon;
+export const UnnnicIcon = icon;
+export const UnnnicIconSvg = icon;
+export const UnnnicIconLoading = iconLoading;
+export const UnnnicToolTip = toolTip;
+export const UnnnicCard = card;
+export const UnnnicSimpleCard = cardSimple;
+export const UnnnicCardCompany = cardCompany;
+export const UnnnicCardData = cardData;
+export const UnnnicCardImage = cardImage as VueComponent;
+// export const UnnnicCardFlow = cardFlow;
+export const UnnnicCardProject = cardProject as VueComponent;
+export const UnnnicCardInformation = cardInformation;
+export const UnnnicCheckbox = checkbox;
+export const UnnnicCollapse = collapse;
+export const UnnnicRadio = radio;
+export const UnnniclanguageSelect = languageSelect as VueComponent;
+export const UnnnicModal = modal;
+export const UnnnicModalDialog = ModalDialog;
+export const UnnnicModalNext = ModalNext;
+export const UnnnicModalUpload = modalUpload;
+export const UnnnicSelectSmart = selectSmart as VueComponent;
+// export const UnnnicSelect = select;
+export const UnnnicSelectItem = selectItem;
+// export const UnnnicSelectListItem = selectListItem;
+export const UnnnicMultiSelect = multiSelect as VueComponent;
+export const UnnnicAlert = alert;
+export const UnnnicCallAlert = (props: any) => callAlert(props);
+export const UnnnicCallModal = (props: any) => callModal(props);
+// export const UnnnicAutocomplete = autocomplete;
+// export const UnnnicAutocompleteSelect = autocompleteSelect;
+export const UnnnicTag = tag;
+export const UnnnicAccordion = accordion;
+export const UnnnicIndicator = indicator;
+export const UnnnicSkeletonLoading = skeletonLoading;
+export const UnnnicCarousel = carousel;
+export const UnnnicLabel = label;
+export const UnnnicTab = tab;
+export const UnnnicTabsExpanded = tabsExpanded;
+export const UnnnicBanner = banner;
+export const UnnnicComment = comment;
+export const UnnnicDatePicker = datePicker;
+export const UnnnicBreadcrumb = breadcrumb;
+export const UnnnicSwitch = Switch;
+export const UnnnicSlider = Slider;
+export const UnnnicDataArea = DataArea;
+export const UnnnicPagination = Pagination;
+export const UnnnicDropArea = DropArea;
+export const UnnnicUploadArea = UploadArea;
+export const UnnnicImportCard = ImportCard;
+export const UnnnicDateFilter = DateFilter;
+export const UnnnicChatText = ChatText;
+export const UnnnicTextArea = TextArea;
+export const UnnnicCardNumber = CardNumber;
+export const UnnnicChartRainbow = chartRainbow;
+export const UnnnicChartBar = chartBar;
+export const UnnnicChartLine = chartLine;
+export const UnnnicMoodRating = moodRating;
+export const UnnnicStarRating = starRating;
+export const UnnnicAudioRecorder = audioRecorder;
+export const UnnnicCircleProgressBar = circleProgressBar;
+export const UnnnicProgressBar = progressBar;
+export const UnnnicChatsContact = ChatsContact;
+export const UnnnicChatsDashboardTagLive = ChatsDashboardTagLive;
+export const UnnnicChatsHeader = ChatsHeader as VueComponent;
+export const UnnnicChatsMessage = ChatsMessage;
+export const UnnnicReplyMessage = ReplyMessage;
+export const UnnnicChatsNavbar = ChatsNavbar;
+export const UnnnicChatsUserAvatar = ChatsUserAvatar;
+export const UnnnicChartMultiLine = ChartMultiLine;
+export const UnnnicEmojiPicker = EmojiPicker;
+export const UnnnicChartFunnel = ChartFunnel as VueComponent;
+export const UnnnicDisclaimer = Disclaimer;
+export const UnnnicDrawer = Drawer;
+export const UnnnicTableNext = TableNext;
+export const UnnnicTour = Tour;
+export const UnnnicNavigator = Navigator;
+export const UnnnicDataTable = DataTable as VueComponent;
+export const UnnnicSelectTime = SelectTime as VueComponent;
+export const UnnnicChip = Chip;
+export const UnnnicTemplatePreview = TemplatePreview as VueComponent;
+export const UnnnicTemplatePreviewModal = TemplatePreviewModal as VueComponent;
